fix: separate logic to use-slider.tsx
